SVS compiles on Windows

+++ src/soar_interface.h
@@ -168,7 +168,7 @@ inline bool soar_interface::get_val(Symbol *sym, double &v) {
 		return true;
 	}
 	if (is_int(sym)) {
-		v = sym->ic.value;
+		v = static_cast<double>(sym->ic.value);
 		return true;
 	}
 	return false;
